refactor: standardize import paths and formatting in NotificationController and SendNotification
