+++ src/server.js
@@ -1,14 +1,13 @@
-// src/server.js
+ // src/server.js
 require("dotenv").config();
 const mongoose = require("mongoose");
+mongoose.set("strictQuery", false); // opcional: suprime warnings futuros de strictQuery
 const app = require("./app");
 
 const PORT = process.env.PORT || 3000;
+
 mongoose
-	.connect(process.env.MONGO_URI, {
-		useNewUrlParser: true,
-		useUnifiedTopology: true,
-	})
+	.connect(process.env.MONGO_URI) // sem flags depreceadas
 	.then(() => {
 		console.log("MongoDB conectado");
 		app.listen(PORT, () => console.log(`Servidor rodando na porta ${PORT}`));
@@ -17,3 +16,4 @@ mongoose
 		console.error("Erro ao conectar no MongoDB:", err);
 		process.exit(1);
 	});
+
